Moved adaptors to their own packages
